rename merge-rules.ts to utils-rules.ts
